Consider which gradients are needed in autograd/find_tape_or_recomp_stmts
