+++ include/autograd/find_tape_or_recomp_stmts.h
@@ -18,6 +18,7 @@ std::pair<std::unordered_map<ID, std::unordered_set<ID>>,
           std::unordered_map<ID, std::unordered_set<ID>>>
 findTapeOrRecompStmts(
     const Stmt &op, const std::unordered_set<ID> &defsToTape,
+    const std::unordered_set<ID> defsNeedGrad,
     std::unordered_map<StmtOrExprID, Derivative::LazyFullDerivative>
         &derivatives);
